BUGFIX: We don't click on all the icon that look like a "save"
Otherwise, we could save two rows at the same time when doing that. As a
result we get an error message for the second one (the new row). It
blocks the tests.

This commit also save the content of the current HTML in the browser
when we wait for too long on a node to appear.