Use bootstrap 3 for the scenario web page
Use glyphicons and not images in the interface.

Add the last error message in the scenario web page.

Add metadata to add a scenario name/description in the output directory

New configuration variable to limit the time we wait on an element
(above that, we look on that as a failure)

New exception to manage the "standard errors" that we predict

"Double login" if a sql cursor is corrupted (we have ot try to log twice
to get rid of it)

New web site to display the results.

New gitignore entry to get rid of the temp directory in the website
package.