Handle multiple tables correctly
When looking for a row on a frame with multiple tables,
if the column was not found in the first table, we'd
get stuck doing retries on that table. Now we do not retry,
so that later tables in the frame can be searched too.

Also: some cleanups proposed by pyflakes