Performance testing cleanup
Get rid of run.sh, which wants to run as root and only
contributed confusion. It might come back when we fix
the Docker images.

Fix export of non-ASCII when converting meta features.

Create directories on demand instead of silently failing.

Use the @testperf tag when doing performance testing to
that the performance tests can be mixed in with the regular
ones.

Add the NORESTORE environment variable, which makes it easier
to test this stuff.

Stop using unbuffer, which is not used in production, and which
apparently does not work with FAKETIME.

Also:

Make the default config.sh usable without running as root.

Save the input to testfield each time we fetch it, so that
we can monitor what tests are changing.